app/requirements.txt
Flask==3.0.0
psycopg2-binary==2.9.9
redis==5.0.1

pytest==8.2.0
ruff==0.6.9
